circles use floating radius

+++ include/gfx.h
@@ -8,7 +8,7 @@ void gfx_clear(u16 col);
 /* draw */
 void gfx_draw_px(int x, int y, u16 col);
 void gfx_draw_line(int x0,int y0,int x1,int y1,u16 c);
-void gfx_draw_circle(int cx, int cy, int r, u16 col);
+void gfx_draw_circle(int cx, int cy, float r, u16 col);
 
 #endif /* GFX_H */
