+++ source/gfx.c
@@ -44,9 +44,9 @@ void gfx_draw_line(int x0,int y0,int x1,int y1,u16 c)
     }
 }
 
-void gfx_draw_circle(int cx, int cy, int r, u16 col)
+void gfx_draw_circle(int cx, int cy, float r, u16 col)
 {
-	int x = r;
+	float x = r;
 	int y = 0;
 
 	int err = 0;
